refactor: :recycle: modularize las rutas /upload y /download
